- Add CloudUtils class, which contains REST functions to interact with cloud server - Refactored Store and Item classes to decouple from User class - Added separate endpoints for saving/updating/restoring store data in cloud server

+++ console/src/test/kotlin/taskquest/console/views/MainTest.kt
@@ -5,11 +5,15 @@ import java.io.File
 
 internal class MainTest {
     private fun mainTestFunc(params: Array<String>) {
-        File("data.json").copyTo(File("backup.json"), true)
-        File("data.json").delete()
+        if (File("data.json").exists()) {
+            File("data.json").copyTo(File("backup.json"), true)
+            File("data.json").delete()
+        }
         main(params)
-        File("backup.json").copyTo(File("data.json"), true)
-        File("backup.json").delete()
+        if (File("backup.json").exists()) {
+            File("backup.json").copyTo(File("data.json"), true)
+            File("backup.json").delete()
+        }
     }
 
     @Test
